Large security commit
After a security assessment we found some worifull things. eg. for pam auth to function properly the server must run as root, this makes some security implications in eg. the command adapter, and the config editor.

This commit changes the configuration a bit, we split the configuration in tree parts, security, authentication and server config.

Security config is read-only config for the system and intoduses ways to  disable adapters and secure who can talk to the different parts of the system. it also made ready to make the command adapter  de-priviledging to a normal user and not start tings as root.

the config editor  have also gotten a touchup so you can make parts of the config read-only.