Refactoring loopback/command data-path.
These port adapters was still using the old port -> listener datapaths with internal queues, these paths are considered legazy and should be removed. as eg. logging is handled trough the new port_manager paths so it is essential that all adapters use this path. this commit moves data to the new port-manager paths, but the old paths are still present so needs to be removed at a later point in time.  i've opened a tidlet in github for this.