Logging data sanitazion and only log inbound direction from port adapter
The logging code had a few issues, for the first it did not sanitize the data inputted. that means the logfile saved all non-visible characters that also made issues showing the files.

it also saved both directions without any indication of the direction, so chars was saved twice.  the code is now changed so it only saves incomming data from the port by default.  (can be changed)

the third change is a switch of in / out directions. i think the direction was out of the port adapter to the client, so out was data recieved on the port and in was from the client. this direction was quite backwards so they are now switched. in = data from port to client  , out = from client to port