Fixes #12: config_editor: Move edit dialogs right under the opened element
the previous solution was a bit counter-intuitive when having long lists with eg. ports.  when pressing edit the editor would open under the table, so pressing the edit button would not show any change befor the user scrolled down the page... this changes the edit dialog to be located right underneath the currently edited item instead.